backend/requirements.txt
fastapi
uvicorn
mangum
python-multipart
numpy
onnxruntime
Pillow